requirements.txt
asgiref==3.8.1
Django==5.0.2
djangorestframework==3.15.1
numpy==2.4.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.4.4
tzdata==2024.1
